requirements.txt
scikit-learn~=1.5.2
numpy~=1.26.4
pillow~=11.0.0
opencv-python~=4.7.0.68
streamlit~=1.17.0
scikit-image~=0.24.0
torch~=2.6.0
matplotlib~=3.10.0
torchvision~=0.21.0
tqdm~=4.67.1
tensorboard
altair==4.0
